+++ tests/KeyPairTest.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * Encryption
+ * Copyright 2020 Jamiel Sharief.
+ *
+ * Licensed under The MIT License
+ * The above copyright notice and this permission notice shall be included in all copies or substantial
+ * portions of the Software.
+ *
+ * @copyright   Copyright (c) Jamiel Sharief
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
+ */
+namespace Encryption\Test;
+
+class KeyPairTest extends \PHPUnit\Framework\TestCase
+{
+    public function testToString()
+    {
+        $this->markTestIncomplete();
+    }
+}