+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2020-11-07
+
+### Added
+
+- Added AsymmetricEncryption::generatePrivateKey
+- Added AsymmetricEncryption::extractPublicKey
 
 ## [0.2.1] - 2020-11-06